+++ lib/init.php
@@ -281,6 +281,58 @@ function mai_save_template_part_delete_transient( $post_id, $post, $update ) {
 	delete_transient( 'mai_template_parts' );
 }
 
+add_action( 'plugins_loaded', 'mai_plugin_update_checker' );
+/**
+ * Initialize plugin update checker.
+ *
+ * @since 0.1.0
+ * @since 2.24.0 Moved to plugins_loaded hook. See #607.
+ *
+ * @return void
+ */
+function mai_plugin_update_checker() {
+	if ( ! current_user_can( 'install_plugins' ) ) {
+		return;
+	}
+
+	if ( ! class_exists( 'Puc_v4_Factory' ) ) {
+		return;
+	}
+
+	$updater = Puc_v4_Factory::buildUpdateChecker(
+		'https://github.com/maithemewp/mai-engine',
+		realpath( __DIR__ . '/..' ) . '/mai-engine.php',
+		'mai-engine'
+	);
+
+	// Get the branch. If checking for beta releases.
+	$branch = function_exists( 'genesis_get_option' ) && genesis_get_option( 'mai_tester' );
+	$branch = $branch ? 'beta' : 'master';
+
+	// Allow branch and updater object manipulation.
+	$branch = apply_filters( 'mai_updater_branch', $branch );
+
+	// Set the branch.
+	$updater->setBranch( $branch );
+
+	// Maybe set github api token.
+	if ( defined( 'MAI_GITHUB_API_TOKEN' ) ) {
+		$updater->setAuthentication( MAI_GITHUB_API_TOKEN );
+	}
+
+	// Add icons for Dashboard > Updates screen.
+	$updater->addResultFilter(
+		function ( $info ) {
+			$info->icons = [
+				'1x' => mai_get_url() . 'assets/img/icon-128x128.png',
+				'2x' => mai_get_url() . 'assets/img/icon-256x256.png',
+			];
+
+			return $info;
+		}
+	);
+}
+
 add_action( 'plugins_loaded', 'mai_load_vendor_plugins' );
 /**
  * Load mai-engine included plugin files.
@@ -447,7 +499,6 @@ function mai_load_files() {
 				'admin/setup-wizard',
 				'admin/term-image',
 				'admin/template-parts',
-				'admin/update-checker',
 				'admin/upgrade',
 			]
 		);